player can save and load game, AB#129
