--- src/SaveGame.java
@@ -1,69 +1,0 @@
-import org.json.simple.parser.ParseException;
-
-import java.io.BufferedReader;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-
-public class SaveGame {
-
-    static Game game;
-    //add extends to every class necessary
-    private static void saveGame() {
-        try {
-            FileOutputStream fos = new FileOutputStream("Adv.txt");
-            ObjectOutputStream oos = new ObjectOutputStream(fos);
-            oos.writeObject(game); // game
-            oos.flush(); // write out any buffered bytes
-            oos.close();
-            System.out.print("Game saved\n");
-        } catch (Exception e) {
-            System.out.print("Can't save data.\n"
-                    + e.getClass() + ": " + e.getMessage() + "\n");
-        }
-    }
-
-    private static void loadGame() {
-        try {
-            FileInputStream fis = new FileInputStream("Adv.txt");
-            ObjectInputStream ois = new ObjectInputStream(fis);
-            game = (Game) ois.readObject();
-            ois.close();
-            System.out.print("\n---Game loaded---\n");
-        } catch (Exception e) {
-            System.out.print("Can't load data.\n");
-            System.out.print(e.getClass() + ": " + e.getMessage() + "\n");
-        }
-    }
-
-    public static void main(String[] args) throws IOException, ParseException {
-        BufferedReader in;
-        String input;
-        String output = "";
-        game = new Game();
-        in = new BufferedReader(new InputStreamReader(System.in));
-        game.showIntro();
-        do {
-            System.out.print("> ");
-            input = in.readLine();
-            switch (input) {
-                case "save":
-                    saveGame();
-                    break;
-                case "load":
-                    loadGame();
-                    break;
-                default:
-                    output = game.runCommand(input);
-                    break;
-            }
-            if (!output.trim().isEmpty()) {
-                game.showStr(output);
-            }
-        } while (!"q".equals(input));
-    }
-
-}
